research/scraper-lambda/requirements.txt
requests
beautifulsoup4
qdrant-client
openai
tiktoken
python-dotenv
pyyaml
boto3